requirements.txt
alembic==1.4.3
autopep8==1.5.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-HTTPAuth==4.1.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Pydantic==0.4.0
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==3.25.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
Pillow==8.3.2
pycodestyle==2.6.0
pydantic==1.6.1
PyJWT==2.1.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-http-client==3.3.1
sendgrid==6.4.7
six==1.15.0
SQLAlchemy==1.3.20
starkbank-ecdsa==1.1.0
StringGenerator==0.4.4
toml==0.10.1
Werkzeug==1.0.1
